+++ Controller/Product/productController.js
@@ -9,7 +9,10 @@ const { appError, notFound } = require("../../Middlewares/appError");
 
 const addProductCtrl = async (req, res, next) => {
   //check Product exits
-  try {
+  try { 
+    if (!req.session.authorized) {
+      return next(appError("Bạn cần đăng nhập", 403));
+    }
     const {
       name,
       description,
@@ -105,7 +108,10 @@ const addProductCtrl = async (req, res, next) => {
 //@access Private/Admin
 
 const getAllProductCtrl = async (req, res, next) => {
-  try {
+  try { 
+    if (!req.session.authorized) {
+      return next(appError("Bạn cần đăng nhập", 403));
+    }
     let productQuery = Product.find();
 
     if (req.query.name) {
@@ -190,7 +196,10 @@ const getAllProductCtrl = async (req, res, next) => {
 //@access Private/Admin
 
 const getProductByIdCtrl = async (req, res, next) => {
-  try {
+  try { 
+    if (!req.session.authorized) {
+      return next(appError("Bạn cần đăng nhập", 403));
+    }
     const product = await Product.findById(req.params.id);
     if (!product) {
       next(appError("Không tìm thấy sản phẩm !", 403));
@@ -210,7 +219,10 @@ const getProductByIdCtrl = async (req, res, next) => {
 //@access Private/Admin
 
 const updateProductCtrl = async (req, res, next) => {
-  try {
+  try { 
+    if (!req.session.authorized) {
+      return next(appError("Bạn cần đăng nhập", 403));
+    }
     const {
       name,
       description,
@@ -262,7 +274,10 @@ const updateProductCtrl = async (req, res, next) => {
 //@access Private/Admin
 
 const deleteProductCtrl = async (req, res, next) => {
-  try {
+  try { 
+    if (!req.session.authorized) {
+      return next(appError("Bạn cần đăng nhập", 403));
+    }
     const product = await Product.findByIdAndDelete(req.params.id);
     if (!product) {
       return next(appError("Sản phẩm không tồn tại", 403));
@@ -282,7 +297,9 @@ const deleteProductCtrl = async (req, res, next) => {
 
 // //upload Photo
 // const uploadPhotoProductCtrl = async (req, res, next) => {
-//   try {
+//   try { if (!req.session.authorized) {
+      return next(appError("Bạn cần đăng nhập", 403));
+    }
 //     //find the product to be update
 //     const product = await Product.findById(req.params.id);
 //     // check if product is found
